Заменили на auth.model.User

+++ constructor/views.py
@@ -25,15 +25,15 @@ def get_answers(query):
 
 @csrf_exempt
 def create_test(request):
-	if request.is_ajax():
-		try:
-			user = User.objects.get(login=request.session.get('user'))
+	login = check_sign_in(request)
+	if login:
+		if request.is_ajax():
 			test = Test(
 				title = request.POST['title'],
 				description = request.POST['description'],
 				helps = str_to_bool(request.POST['helps']),
 				time_completion = str_to_bool(request.POST['timeCompl']),
-				creator = user,
+				creator = auth.get_user(request),
 				two_mark = request.POST['two_mark'],
 				three_mark = request.POST['three_mark'],
 				four_mark = request.POST['four_mark']
@@ -41,19 +41,15 @@ def create_test(request):
 			test.save()
 
 			return JsonResponse({'testID': test.id, 'error': False})
-		except ObjectDoesNotExist:
-			return JsonResponse({'error': True})
-	elif 'user' in request.session:
-		sign_in = check_sign_in(request)
-		return render_to_response('create_test.html', {'login': sign_in})
+		else:
+			return render_to_response('create_test.html', {'login': login})
 	else:
 		return redirect('/')
 
 @csrf_exempt
 def settings_test(request, id):
 	test = get_object_or_404(Test, id=id)
-	login = check_sign_in(request)
-	if test.creator == get_object_or_404(User, login=login):
+	if test.creator == auth.get_user(request):
 		if request.is_ajax():
 
 			test.title = request.POST['title']
@@ -69,6 +65,7 @@ def settings_test(request, id):
 
 			return JsonResponse({'success': 'Данные сохранены!'})
 		else:
+			login = check_sign_in(request)
 			return render_to_response('create_test.html', {'login': login, 'test': test})
 	else:
 		return redirect('/')
@@ -76,52 +73,53 @@ def settings_test(request, id):
 @csrf_exempt
 def queries_test(request, id):
 	sign_in = check_sign_in(request)
-	test = Test.objects.get(id=id)
-	queries = map(get_answers, Query.objects.filter(test = test))
-
-	return render_to_response('test/queries.html', {'login': sign_in, 'test': test, 'queries': queries})
+	test = get_object_or_404(Test, id=id)
+	if test.creator == auth.get_user(request):
+		queries = map(get_answers, Query.objects.filter(test = test))
+		return render_to_response('test/queries.html', {'login': sign_in, 'test': test, 'queries': queries})
+	else:
+		raise Http404('Вы не являетесь создателем данного теста!')
 
 @csrf_exempt
 def add_query(request, id):
-	login = check_sign_in(request)
-	if login:
-		try:
-			test = Test.objects.get(id=int(id))
-			if test.creator.login == login:
-				#кодим тут (когда все True)
-				if request.is_ajax():
-					query = Query(
-						test = test,
-						text = request.POST['text'],
-						point = request.POST['point'],
-						help = request.POST['help'],
-						time = request.POST['time']
-					)
+	test = get_object_or_404(Test, id=int(id))
+	user = auth.get_user(request)
+	if test.creator == user:
+		#кодим тут (когда все True)
+		if request.is_ajax():
+			query = Query(
+				test = test,
+				text = request.POST['text'],
+				point = request.POST['point'],
+				help = request.POST['help'],
+				time = request.POST['time']
+			)
 
-					query.save()
-					for data_answer in json.loads(request.POST['answers']):
-						answer = Answer(
-							query = query,
-							text = data_answer['text'],
-							correct = data_answer['correct']
-						)
+			query.save()
+			for data_answer in json.loads(request.POST['answers']):
+				answer = Answer(
+					query = query,
+					text = data_answer['text'],
+					correct = data_answer['correct']
+				)
 
-						answer.save()
+				answer.save()
 
-					return JsonResponse({'complite': True})
-				else:	
-					return render_to_response('add_query.html', {'login': login, 'test': test})
-			else:
-				raise Http404('Вы не являетесь создателем данного теста!')
-		except:
-			raise Http404('Теста с таким id не существует!')
+			return JsonResponse({'complite': True})
+		else:	
+			return render_to_response('add_query.html', {'login': user, 'test': test})
 	else:
-		return redirect('/')
+		raise Http404('Вы не являетесь создателем данного теста!')
 
 @csrf_exempt
 def delete_query(request, t_id, q_id):
-	Query.objects.get(test=t_id, id=q_id).delete()
-	return redirect('/constructor/test/' + t_id + '/questions/')
+	test = get_object_or_404(Test, id = t_id)
+	if test.creator == auth.get_user(request):
+		question = get_object_or_404(Query, test=test, id=q_id)
+		question.delete()
+		return redirect('/constructor/test/' + t_id + '/questions/')
+	else:
+		raise Http404('Вы не являетесь создателем данного теста!')
 
 @csrf_exempt
 def edit_question(request, t_id, q_id):
@@ -129,7 +127,7 @@ def edit_question(request, t_id, q_id):
 	question = get_object_or_404(Query, id=q_id)
 	login = check_sign_in(request)
 
-	if test.creator.login == login:
+	if test.creator == auth.get_user(request):
 		if request.method == 'POST':
 			#отправляем model
 			return HttpResponse(json.dumps(model_to_dict(question)))
@@ -157,38 +155,42 @@ def edit_question(request, t_id, q_id):
 
 @csrf_exempt
 def question_actions(request, qid, aid):
-	if request.method == 'DELETE':
-		answer = get_object_or_404(Answer, query=qid, id=aid)
-		answer.delete()
-		return JsonResponse({'msg':'delete'})
-	elif request.method == 'PUT':
-		answer = get_object_or_404(Answer, query=qid, id=aid)
-		#парсим put запрос
-		data = put(request)
-		#обновляем данные
-		answer.text = data.get('text')
-		answer.correct = data.get('correct')
-		answer.save()
+	question = get_object_or_404(Query, id=qid)
+	if question.test.creator == auth.get_user(request):
+		if request.method == 'DELETE':
+			answer = get_object_or_404(Answer, query=qid, id=aid)
+			answer.delete()
+			return JsonResponse({'msg':'delete'})
+		elif request.method == 'PUT':
+			answer = get_object_or_404(Answer, query=qid, id=aid)
+			#парсим put запрос
+			data = put(request)
+			#обновляем данные
+			answer.text = data.get('text')
+			answer.correct = data.get('correct')
+			answer.save()
 
-		return JsonResponse({'msg':'success'})
-	elif request.method == 'POST':
-		question = get_object_or_404(Query, id=qid)
-		answer = Answer(query=question)
+			return JsonResponse({'msg':'success'})
+		elif request.method == 'POST':
+			question = get_object_or_404(Query, id=qid)
+			answer = Answer(query=question)
 
-		answer.save()
-		return JsonResponse({'id': answer.id})
+			answer.save()
+			return JsonResponse({'id': answer.id})
 
-	#достаем коллекцию овтетов данного вопроса
-	elif request.method == 'GET':
-		question = get_object_or_404(Query, id=qid)
-		answers = Answer.objects.filter(query=question)
-		response = []
-		#получаем поля модели в json формате и сразу парсим
-		answers = json.loads(serializers.serialize('json', answers))
-		#формируем модели
-		for answer in answers:
-			data = answer.get('fields')
-			data['id'] = answer.get('pk')
-			response.append(data)
-		#возвращаем модели в JSON формате
-		return HttpResponse(json.dumps(response))
+		#достаем коллекцию овтетов данного вопроса
+		elif request.method == 'GET':
+			question = get_object_or_404(Query, id=qid)
+			answers = Answer.objects.filter(query=question)
+			response = []
+			#получаем поля модели в json формате и сразу парсим
+			answers = json.loads(serializers.serialize('json', answers))
+			#формируем модели
+			for answer in answers:
+				data = answer.get('fields')
+				data['id'] = answer.get('pk')
+				response.append(data)
+			#возвращаем модели в JSON формате
+			return HttpResponse(json.dumps(response))
+	else:
+		return Http404('Вы не являетесь создателем данного теста!')
